fix(db/dsm): 修正MySQL下id字段类型为INTEGER
将原本MySQL下id字段使用的BIGINT改为INTEGER，统一不同数据库的id字段类型配置逻辑

+++ server/src/db/dsm/schema-reconciler.ts
@@ -809,7 +809,7 @@ export class SchemaReconciler {
 
   private mapTypeToSQL(type: string, dbType: string, length?: number): string {
     switch (type) {
-      case 'id': return dbType === 'postgresql' ? 'SERIAL' : (dbType === 'mysql' ? 'BIGINT' : 'INTEGER');
+      case 'id': return dbType === 'postgresql' ? 'SERIAL' : (dbType === 'mysql' ? 'INTEGER' : 'INTEGER');
       case 'number': return dbType === 'postgresql' ? 'BIGINT' : 'BIGINT';
       case 'integer': return 'INTEGER';
       case 'boolean': return dbType === 'postgresql' ? 'BOOLEAN' : (dbType === 'mysql' ? 'TINYINT(1)' : 'INTEGER');
